Mapea resultados (Map results)

Función para mapear resultados de la elección (Function to map election results)

map_results(data)

Arguments

data

un tibble guardado como objeto en el Enviroment luego de consultar get_election_data con parámetro level en provincia para elecciones presidenciales y level en departamento para legislativas (tibble saved as an object in the Enviroment after querying get_election_data with provincia as level parameter for presidential elections and departmanto for legislative elections). #'

Value

Devuelve un objeto de class"gg" "ggplot" que mapea el resultado de una eleccion condicional al nivel de agregacion de data (Returns an object of class "gg" "ggplot" that plots the election results in a map conditional on the level of data aggregation.).

Details

REQUISITOS:

1. El formato de data debe ser long para poder graficar. Si data es wide se puede transformar con make_long (long format of data is required for plotting results. If data is in wide format you can transform it with make_long)

2. data tiene que haber incorporando los nombres de las listas. Agreguelos con electorAr::get_names (data must have party names. Add them with electorAr::get_names)

See also